+++ src/sso.js
@@ -3,9 +3,10 @@
  */
 
 import ethersjs from 'ethers'
-import { isEmpty, isSet } from './utils/helpers.js'
+import { currentTimestamp, isEmpty, isEthereumAddress } from './utils/helpers.js'
 import jwt from 'jsonwebtoken'
 import { toJwtPayload } from './models/index.js'
+import { createProvider, Registry } from './contracts/index.js'
 
 const { ethers } = ethersjs
 
@@ -45,11 +46,11 @@ export const messageToSign = (data = {}) => {
  * await generateSSORequestParams(domainOwnerPrivateKey, {redirectUrl: 'http://silkey.io', refId: 1});
  */
 export const generateSSORequestParams = async (privateKey, data = {}) => {
-  const redirectUrl = data.redirectUrl || null
-  const cancelUrl = data.cancelUrl || null
-  const sigTimestamp = data.sigTimestamp || Math.round(Date.now() / 1000)
-  const refId = data.refId || null
-  const scope = data.scope || null
+  const redirectUrl = data.redirectUrl || ''
+  const cancelUrl = data.cancelUrl || ''
+  const sigTimestamp = data.sigTimestamp || currentTimestamp()
+  const refId = data.refId || ''
+  const scope = data.scope || ''
 
   wallet = new ethers.Wallet(privateKey)
 
@@ -78,14 +79,17 @@ export const verifyUserSignature = tokenPayload => {
   try {
     const payload = toJwtPayload(tokenPayload)
 
-    if (isEmpty(payload.userSignature) || isEmpty(payload.address)) {
-      console.warn('Verification failed, missing user signature and/or address')
+    if (isEmpty(payload.userSignature) || isEmpty(payload.userSignatureTimestamp) || isEmpty(payload.address)) {
+      console.warn('Verification failed, missing user signature/timestamp and/or address')
       return false
     }
 
     const message = payload.messageToSignByUser()
     const signer = ethers.utils.verifyMessage(message, payload.userSignature)
-    return signer.toLowerCase() === payload.address.toLowerCase()
+    const success = signer.toLowerCase() === payload.address.toLowerCase()
+    !success && console.warn(`verifyUserSignature: expect ${signer} to be equal ${payload.address}`)
+
+    return success
   } catch (e) {
     console.warn(e)
     return false
@@ -101,33 +105,55 @@ export const verifyUserSignature = tokenPayload => {
  * @param silkeyPublicKey {string} optional
  * @return {null|boolean}
  */
-export const verifySilkeySignature = (tokenPayload, silkeyPublicKey = null) => {
+export const verifySilkeySignature = (tokenPayload, silkeyPublicKey) => {
   try {
+    const payload = toJwtPayload(tokenPayload)
+
+    if (isEmpty(payload.email) && isEmpty(payload.silkeySignature)) {
+      return null
+    }
+
+    if (isEmpty(payload.email) ^ isEmpty(payload.silkeySignature)) {
+      console.warn('Verification failed, missing silkey signature or email')
+      return false
+    }
+
+    if (isEmpty(payload.silkeySignatureTimestamp)) {
+      console.warn('Verification failed, missing silkey signature timestamp')
+      return false
+    }
+
+    const signer = ethers.utils.verifyMessage(payload.messageToSignBySilkey(), payload.silkeySignature)
+
     if (isEmpty(silkeyPublicKey)) {
       console.warn('You are using verification without checking silkey signature. We strongly recommended to turn on full verification. This option can be deprecated in the future')
       return true
     }
 
-    const payload = toJwtPayload(tokenPayload)
+    const success = signer.toLowerCase() === silkeyPublicKey.toLowerCase()
+    !success && console.warn(`verifySilkeySignature: expect ${signer} to be equal ${silkeyPublicKey}`)
 
-    if (isSet(payload.email) ^ isSet(payload.silkeySignature)) {
-      console.warn('Verification failed, missing silkey signature or email')
-      return false
-    }
-
-    if (!isSet(payload.email) && !isSet(payload.silkeySignature)) {
-      return null
-    }
-
-    const signer = ethers.utils.verifyMessage(payload.messageToSignBySilkey(), payload.silkeySignature)
-    const result = signer.toLowerCase() === silkeyPublicKey.toLowerCase()
-    !result && console.warn(`verifySilkeySignature: expect ${signer} to be equal ${silkeyPublicKey}`)
-
-    return result
+    return success
   } catch (e) {
     console.warn(e)
     return false
   }
+}
+
+/**
+ * Fetches public ethereum Silkey address directly from blockchain
+ *
+ * @param providerUri {string} ie: 'https://infura.io/v3/:infuraId' register to infura.io to get infuraId
+ * @param registryAddress {string} address of silkey smart contract registry, see list of addresses in README#registryAddress
+ * @return {Promise<string>} public ethereum address of silkey signer
+ */
+export const fetchSilkeyPublicKey = async (providerUri, registryAddress) => {
+  const provider = createProvider(providerUri)
+  const registry = new Registry(provider, registryAddress)
+  const key = await registry.getAddress('Hades')
+
+  if (!isEthereumAddress(key)) throw Error(`Invalid key: ${key}`)
+  return key
 }
 
 /**
@@ -141,7 +167,7 @@ export const verifySilkeySignature = (tokenPayload, silkeyPublicKey = null) => {
  * @throws when token is invalid or data are corrupted
  * @example
  * // returns {JwtPayload}
- * tokenPayloadVerifier('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6IjB4M0EzQTYyMzUxRkFlN2QxOTNlZjUxNTU3MUIxRjM1OTUwNzhEZDllMEBwcml2YXRlcmVsYXkuc2lsa2V5LmlvIiwiYWRkcmVzcyI6IjB4M0EzQTYyMzUxRkFlN2QxOTNlZjUxNTU3MUIxRjM1OTUwNzhEZDllMCIsInNpZ25hdHVyZSI6eyJyIjoiMHg5NzhmZTdhZmMwODY1NTk4YTNiYTNmOGUzMTI0ZDBkMGM3MGYyMTMwOTQ5YTBhZDRiZTk3ODc5MWI0ZGQ2Y2Q3IiwicyI6IjB4MDU1NjEwZGYzZmI2ODAyYzgwZjQ0NzVjNjIyNDc1OGM0Y2VjNWVkMTllMTMzN2YwODEwMmM3NjNlYWM2Y2JjMyIsIl92cyI6IjB4ODU1NjEwZGYzZmI2ODAyYzgwZjQ0NzVjNjIyNDc1OGM0Y2VjNWVkMTllMTMzN2YwODEwMmM3NjNlYWM2Y2JjMyIsInJlY292ZXJ5UGFyYW0iOjEsInYiOjI4fSwiaWF0IjoxNjAyMTQ0OTk2fQ.eU3B-jHnu8ToKeU9833jhr9Klvzwpb_oY60Q_jDW0js');
+ * tokenPayloadVerifier('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c');
  */
 export const tokenPayloadVerifier = (token, silkeyPublicKey = null) => {
   try {