+++ test/keys.js
@@ -1,0 +1,3 @@
+export const publicKey = '0xDBF03b99664deb3C73045ac8933A6db89fefFf5F'
+// this is random PK, please do not use it for anything else than tests
+export const privateKey = '0x2c06e0037dacc4a831049ce0770f5f6f788827659a5842ed96d34c0631d5f6de'